requirements.txt
motor~=3.1.2
aiogram~=2.25.1
python-dateutil~=2.8.2